fresh news

+++ smarti-cloud/src/main/java/com/filemark/jcr/controller/SiteController.java
@@ -1169,13 +1169,13 @@ public class SiteController extends BaseController {
 	}
     	
 	@RequestMapping(value = {"/site/getassets.json"}, method = RequestMethod.GET)
-	public @ResponseBody Map<String, News[]> getAssetsJson(String path,String type,Integer w,Model model,HttpServletRequest request, HttpServletResponse response){
+	public @ResponseBody Map<String, News[]> getAssetsJson(String path,String type,String r,Integer w,Model model,HttpServletRequest request, HttpServletResponse response){
 		Folder folder = folderJson(path,type,model,request,response);
 		if(w ==null) w = 4; 
 		Map<String, News[]> data = new HashMap<String, News[]>();	
 		File json = new File(jcrService.getDevice()+path+"/Assets_"+type+"_"+w+".json");
 		try {	
-			if(json.exists() && folder.getLastModified() !=null && json.lastModified() > folder.getLastModified().getTime()) {
+			if(r==null && json.exists() && folder.getLastModified() !=null && json.lastModified() > folder.getLastModified().getTime()) {
 				response.setContentType("application/json");
 				FileUtils.copyFile(json, response.getOutputStream());
 				return null;
@@ -1228,6 +1228,12 @@ public class SiteController extends BaseController {
 						+"<video poster=\"video2jpg.jpg?path="+asset.getPath()+"\" width=\"400\" height=\"300\" controls=\"controls\"  preload=\"none\">"
 						+"<source type=\"video/mp4\" src=\"video.mp4?path="+asset.getPath()+"\"/></video></figture>";
 			}
+	        if(asset.getDoc2pdf()) {
+	        	title = "<a class=\"download pull-right\" href=\"file/"+asset.getName()+"?path=\""+asset.getPath()+" target=\"_BLANK\" download><span class=\"glyphicon glyphicon-download\">下载</span></a>"
+			    		+"<a class=\""+asset.getCssClass()+"\" href=\"doc2pdf.pdf?path="+asset.getPath()+" target=\"_BLANK\">"
+					    +"<img id=\"img"+asset.getUid()+"\" src=\""+icon+"\" class=\"img-responsive img-rounded pull-left mrgn-rght-md\" draggable=\"true\"/>"
+					    +asset.getTitle()+"</a>";
+	        }			
 			a2news.setTitle(title);
 			a2news.setDescription(asset.getDescription()==null?"":asset.getDescription());
 			if(asset.getOriginalDate()!=null)
